+++ src/robot_dart/graphics/base_graphics.hpp
@@ -1,0 +1,28 @@
+#ifndef ROBOT_DART_GRAPHICS_BASE_GRAPHICS_HPP
+#define ROBOT_DART_GRAPHICS_BASE_GRAPHICS_HPP
+
+#include <dart/simulation/World.hpp>
+
+namespace robot_dart {
+    namespace graphics {
+        class BaseGraphics {
+        public:
+            BaseGraphics() {}
+
+            BaseGraphics(dart::simulation::WorldPtr world) {}
+
+            virtual bool done() const
+            {
+                return false;
+            }
+
+            virtual void refresh() {}
+
+            virtual void set_render_period(double dt) {}
+
+            virtual void set_enable(bool enable) {}
+        };
+    } // namespace graphics
+} // namespace robot_dart
+
+#endif
